Move components to shared folder
