+++ src/components/PatchesPage/patchCard/BuildStatusIcon.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import { Tooltip } from "antd";
 import { Link } from "react-router-dom";
 import Icon from "components/icons/Icon";
-import { CreatedIcon } from "pages/userPatches/patchCard/BuildStatusIcon/CreatedIcon";
 import { BuildStatus } from "types/build";
+import { CreatedIcon } from "./buildStatusIcon/CreatedIcon";
 
 interface Props {
   status: string;
